Unfork query.py
